Initial YDB support
Combined commits:
c8bb408273f move ydb to pkg, fix replace
f006532183e Merge pull request #34 from ydb-platform/logging
890b82f3878 added logging
5c68c7e4d55 formatting
593aa1c5570 make gofmt
2398f13cfc8 Merge pull request #33 from ydb-platform/fix-check-err-retry
d3a526eff31 simplify check err is IsRetryable
a8fdd476de7 Initial YDB support

+++ pkg/services/playlist/playlistimpl/xorm_store.go
@@ -36,7 +36,7 @@ func (s *sqlStore) Insert(ctx context.Context, cmd *playlist.CreatePlaylistComma
 			return err
 		}
 		if count > MAX_PLAYLISTS {
-			return fmt.Errorf("too many playlists exist (%d > %d)", count, MAX_PLAYLISTS)
+			return fmt.Errorf("too many playlists exist (%d > %d)", count, uint64(MAX_PLAYLISTS))
 		}
 
 		ts := time.Now().UnixMilli()
@@ -199,7 +199,7 @@ func (s *sqlStore) ListAll(ctx context.Context, orgId int64) ([]playlist.Playlis
 	db := s.db.GetSqlxSession() // OK because dates are numbers!
 
 	playlists := []playlist.PlaylistDTO{}
-	err := db.Select(ctx, &playlists, "SELECT * FROM playlist WHERE org_id=? ORDER BY created_at asc LIMIT ?", orgId, MAX_PLAYLISTS)
+	err := db.Select(ctx, &playlists, "SELECT * FROM playlist WHERE org_id=? ORDER BY created_at asc LIMIT ?", orgId, uint64(MAX_PLAYLISTS))
 	if err != nil {
 		return nil, err
 	}
